c++: Refer to internal linkage for -Wsubobject-linkage [PR86491]
Since C++11 relaxed the requirement for template arguments to have
external linkage, it's possible to get -Wsubobject-linkage warnings
without using any anonymous namespaces. This confuses users when they
get diagnostics that refer to an anonymous namespace that doesn't exist
in their code.

This changes the diagnostic to say "has internal linkage" for C++11 and
later, which is accurate whether internal linkage is due to the 'static'
specifier, or due to the use of anonymous namespaces.

For C++98 template arguments declared with 'static' are ill-formed
anyway, so the only way this warning can arise is via anonymous
namespaces. That means the existing wording is accurate for C++98 and so
we can keep it.

	PR c++/86491

gcc/cp/ChangeLog:

	* decl2.cc (constrain_class_visibility): Adjust wording of
	-Wsubobject-linkage to account for cases where anonymous
	namespaces aren't used.

gcc/testsuite/ChangeLog:

	* g++.dg/warn/Wsubobject-linkage-3.C: Adjust for new warning.
	* g++.dg/warn/anonymous-namespace-1.C: Use separate dg-warning
	directives for C++98 and everything else.
	* g++.dg/warn/anonymous-namespace-2.C: Likewise.
	* g++.dg/warn/anonymous-namespace-3.C: Likewise.