c++: Define built-in for std::tuple_element [PR100157]
Add __builtin_type_pack_element so that std::tuple_element and
std::variant_alternative can be implemented efficiently, without
recursive class template instantiations.

The name is intended to be similar to Clang's __type_pack_element<>
built-in which has the same behaviour, but uses template syntax instead
of function-call syntax. As the syntax isn't the same, using an
identical name isn't an option. Instead the name is "as close as
possible, but no closer".

The libstdc++ headers can be updated to use this new built-in, or
Clang's equivalent when available. The _Nth_type_t alias template is
defined to use the built-in directly to avoid even instantiating
_Nth_type, and most places in the library headers now use the alias.
There's one place in <variant> that still needs to use the _Nth_type
class template, because the result appears in mangled names and the
built-in can't be mangled. As noted in a comment, this is a feature not
a bug, as it means we always mangle uses of _Nth_type consistently,
rather than depending on how it's implemented (i.e. with GCC's built-in,
Clang's built-in, or pure C++).

	PR c++/100157

gcc/c-family/ChangeLog:

	* c-common.cc (c_common_reswords): Add
	__builtin_type_pack_element.
	* c-common.h (enum rid): Add RID_TYPE_PACK_ELEMENT.

gcc/cp/ChangeLog:

	* constraint.cc (diagnose_trait_expr): Add
	CPTK_TYPE_PACK_ELEMENT to switch.
	* cp-objcp-common.cc (names_builtin_p): Add
	RID_TYPE_PACK_ELEMENT to switch.
	(cp_common_init_ts): Mark TYPE_PACK_ELEMENT as not having a
	common member.
	* cp-tree.def (TYPE_PACK_ELEMENT): Define tree code.
	* cp-tree.h (enum cp_trait_kind): Add CPTK_TYPE_PACK_ELEMENT.
	(TYPE_PACK_ELEMENT_ARGS): Define.
	(finish_type_pack_element): Declare.
	* error.cc (dump_type): Add TYPE_PACK_ELEMENT to switch.
	(dump_type_prefix): Likewise.
	(dump_type_suffix): Likewise.
	* mangle.cc (write_type): Likewise.
	* parser.cc (cp_keyword_starts_decl_specifier_p): Add
	RID_TYPE_PACK_ELEMENT to switch.
	(cp_parser_trait_expr): Likewise. Parse its arguments and call
	finish_type_pack_element.
	(cp_parser_simple_type_specifier): Add RID_TYPE_PACK_ELEMENT to
	switch.
	* pt.cc (for_each_template_parm_r): Add TYPE_PACK_ELEMENT to
	switch.
	(tsubst): Likewise.
	(unify): Likewise.
	(dependent_type_p_r): A TYPE_PACK_ELEMENT is dependent.
	* semantics.cc (finish_type_pack_element): New function.

gcc/ChangeLog:

	* doc/extend.texi (Type Traits): Document new built-in.

libstdc++-v3/ChangeLog:

	* include/bits/utility.h (_Nth_type_t): New alias template using
	built-ins when available.
	* include/std/tuple: Use _Nth_type_t instead of _Nth_type.
	* include/std/variant: Likewise.
	* testsuite/20_util/tuple/element_access/get_neg.cc: Prune
	additional errors from the new built-in.

gcc/testsuite/ChangeLog:

	* g++.dg/ext/type_pack_element1.C: New test.
	* g++.dg/ext/type_pack_element2.C: New test.